Added implementation for Mock Service and Modified api-spec
