+++ src/main/java/io/todo/tasksms/TasksMsApplication.java
@@ -1,9 +1,10 @@
-package net.todo.tasksms;
+package io.todo.tasksms;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 public class TasksMsApplication {
 
 	public static void main(String[] args) {
