+++ src/main/resources/application.yaml
@@ -1,0 +1,4 @@
+spring:
+  data:
+    mongodb:
+      uri: mongodb+srv://admin:root@todo-cluster-0.nscq6.mongodb.net/task?retryWrites=true&w=majority
