--- src/main/resources/application.yml
@@ -1,4 +1,0 @@
-spring:
-  data:
-    mongodb:
-      uri: mongodb+srv://admin:root@todo-cluster-0.nscq6.mongodb.net/task?retryWrites=true&w=majority
