+++ src/test/java/io/todo/tasksms/TasksMsApplicationTests.java
@@ -1,4 +1,4 @@
-package net.todo.tasksms;
+package io.todo.tasksms;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
